requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
numpy
python-dotenv
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
duckduckgo-search
tensorflow
scikit-learn
sentence-transformers
chromadb
watchfiles
